+++ cmd/validator/exit/process.go
@@ -614,7 +614,7 @@ func (c *command) generateDomain(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
-	forkVersion, err := c.obtainForkVersion(ctx)
+	forkVersion, err := c.obtainExitForkVersion(ctx)
 	if err != nil {
 		return err
 	}
@@ -661,10 +661,11 @@ func (c *command) obtainGenesisValidatorsRoot(_ context.Context) (phase0.Root, e
 	if c.debug {
 		fmt.Fprintf(os.Stderr, "Using genesis validators root %#x\n", genesisValidatorsRoot)
 	}
+
 	return genesisValidatorsRoot, nil
 }
 
-func (c *command) obtainForkVersion(_ context.Context) (phase0.Version, error) {
+func (c *command) obtainExitForkVersion(_ context.Context) (phase0.Version, error) {
 	forkVersion := phase0.Version{}
 
 	if c.forkVersion != "" {
@@ -683,12 +684,13 @@ func (c *command) obtainForkVersion(_ context.Context) (phase0.Version, error) {
 		if c.debug {
 			fmt.Fprintf(os.Stderr, "Fork version obtained from chain info\n")
 		}
-		// Use the current fork version for generating an exit as per the spec.
-		copy(forkVersion[:], c.chainInfo.CurrentForkVersion[:])
+		// Use the Capella fork version for generating an exit as per the spec.
+		copy(forkVersion[:], c.chainInfo.ExitForkVersion[:])
 	}
 
 	if c.debug {
 		fmt.Fprintf(os.Stderr, "Using fork version %#x\n", forkVersion)
 	}
+
 	return forkVersion, nil
 }